separated emulator initialisation from main
