+++ src/init.hpp
@@ -1,0 +1,6 @@
+#ifndef INIT
+#define INIT
+
+extern int init();
+
+#endif
